Refactor import statements to use the dataset module for Dataset class
